pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "cssutils>=2.11.1",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "trafilatura>=2.0.0",
]
